Add 'dotenv' for mongo configuration
Add 'dotenv' dependency, which allows environment variables to be
specified in the '.env' file.

This is useful in this case because it allows custom configuration of
username and password for each developer, without posting credential
information to GitLab.

Dotenv will likely also come in handy later for specifying ports, etc.
on a production server.

Add the 'config/config.js' file, which reads the envoronment variables
specifying MongoDB credentials, and makes them available for use.

Update README to explain how to use `npm install`, as well as how to
configure MongoDB credentials in the `.env` file.